fix: more coverage for parser
