+++ playa/parser.py
@@ -431,21 +431,18 @@ class ObjectParser:
 
     def get_object_reference(self, pos: int, token: Token) -> Union[ObjRef, None]:
         """Get an indirect object reference upon finding an "R" token."""
-        try:
-            _pos, _genno = self.stack.pop()
-            _pos, objid = self.stack.pop()
-        except ValueError as e:
+        _pos, genno = self.stack.pop()
+        _pos, objid = self.stack.pop()
+        if not isinstance(objid, int):
             if self.strict:
                 raise PDFSyntaxError(
-                    "Expected generation and object id in indirect object reference"
-                ) from e
-            else:
-                log.warning(
-                    "Expected generation and object id in indirect object reference: %s",
-                    e,
+                    f"Expected object number and generation id, got {objid!r} {genno!r}"
                 )
+            log.warning(
+                "Expected object number and generation id, got %r %r",
+                objid, genno
+            )
             return None
-        objid = int_value(objid)
         if objid == 0:
             if self.strict:
                 raise PDFSyntaxError(
@@ -523,9 +520,9 @@ class ObjectParser:
         idpos = pos
         (pos, objs) = self.pop_to(KEYWORD_BI)
         if len(objs) % 2 != 0:
-            error_msg = f"Invalid dictionary construct: {objs!r}"
+            error_msg = f"Dictionary contains odd number of objects: {objs!r}"
             if self.strict:
-                raise TypeError(error_msg)
+                raise PDFSyntaxError(error_msg)
             else:
                 log.warning(error_msg)
         dic = {literal_name(k): v for (k, v) in choplist(2, objs) if v is not None}
